requirements.txt
setuptools
wheel
numpy
pandas
ipython
jupyterlab
Pillow
opencv-python
pytesseract
pathlib
pyenchant
gensim
top2vec
nltk
spacy